requirements.txt
pyserial
geolocator